koopajr: the whole TU from the map, with all eight nerves and the UNUSED helpers
TKoopaJr, TKoopaJrSubmarine, TCallbackHitActor, TDirectionCalc, both
params classes and both managers, in the map's order. 41 of 82 functions
match; the rest are mostly frame-size or inlining differences.

Along the way: JGeometry::TUtil<f32>::mod gets the body the binary has
(the same one std::fmodf carries as a weak copy), JGeometry::min joins
max, and TBathtub's killer counts return int so callers stop masking.

+++ include/MoveBG/MapObjCorona.hpp
@@ -41,9 +41,9 @@ public:
 	u8 getNextGrip(const JGeometry::TVec3<f32>&, const JGeometry::TVec3<f32>&,
 	               f32, f32*) const;
 	void updatePosture_();
-	u8 getNumKillerLaunchable() const;
+	int getNumKillerLaunchable() const;
 	bool isKillerAttackable() const;
-	u8 getNumKillerBurstable() const;
+	int getNumKillerBurstable() const;
 	bool isBreaking() const;                                // Unused
 	bool isKillerLaunchable() const;                        // Unused
 	void showMessage(u32);                                  // Unused
